share argument verification across kqlHasAnyIpv4.cpp and kqlHasAnyIpv4Prefix.cpp
