+++ src/Functions/Kusto/kqlHasAnyIpv4.cpp
@@ -7,24 +7,12 @@
 #include <Functions/FunctionHelpers.h>
 #include <Functions/IFunction.h>
 #include <Functions/Kusto/KqlFunctionBase.h>
+#include <Functions/Kusto/kqlHasAnyIp.h>
 #include <regex>
 #include <ranges>
 
 namespace DB
 {
-namespace ErrorCodes
-{
-    extern const int ILLEGAL_TYPE_OF_ARGUMENT;
-    extern const int BAD_ARGUMENTS;
-    extern const int NUMBER_OF_ARGUMENTS_DOESNT_MATCH;
-}
-
-enum ArgumentPolicy
-{
-    AP_Single,
-    AP_Variadic
-};
-
 template <typename Name, ArgumentPolicy ap>
 class FunctionKqlHasIpv4Generic : public IFunction
 {
@@ -43,39 +31,7 @@ public:
 
     DataTypePtr getReturnTypeImpl(const ColumnsWithTypeAndName & arguments) const override
     {
-        const auto args_length = arguments.size();
-
-        if (args_length < 2)
-        {
-            throw Exception(ErrorCodes::NUMBER_OF_ARGUMENTS_DOESNT_MATCH, "Number of arguments for function {} doesn't match: passed {}, should be 2 or more.", getName(), toString(arguments.size()));
-        }
-
-        if (!isStringOrFixedString(arguments.at(0).type))
-        {
-            throw Exception(ErrorCodes::ILLEGAL_TYPE_OF_ARGUMENT, "Illegal type of argument of function {}", getName());
-        }
-
-        if (!isStringOrFixedString(arguments.at(1).type) && !isArray(arguments.at(1).type))
-        {
-            throw Exception(ErrorCodes::ILLEGAL_TYPE_OF_ARGUMENT, "Illegal type of argument of function {}", getName());
-        }
-
-        if (isStringOrFixedString(arguments.at(1).type))
-        {
-            if constexpr (ap == AP_Variadic)
-            {
-                const auto are_arguments_valid = std::ranges::all_of(arguments | std::views::drop(2), [](const auto & argument) { return isStringOrFixedString(argument.type); });
-                if (!are_arguments_valid)
-                    throw Exception(ErrorCodes::ILLEGAL_TYPE_OF_ARGUMENT, "Illegal type of argument of function {}", getName());
-            }
-        }
-
-        else if (ap == AP_Single || !isArray(arguments.at(1).type))
-        {
-            throw Exception(ErrorCodes::BAD_ARGUMENTS, "Illegal type of argument of function {}", getName());
-        }
-
-        return std::make_shared<DataTypeUInt8>();
+        return KqlHasAnyIpArgument::ArgumentVerifier(arguments, getName(), ap);
     }
 
     ColumnPtr executeImpl(
